// packages/api-client/src/generated/models/ContentViewType.ts
/* tslint:disable */
/**
 * Yumpoo API
 * Yumpoo 一期 HTTP API 的唯一静态契约源。  所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共 schema、header、parameter、response 和 golden example；业务 paths 在对应模块 实现时按兼容新增原则加入。
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @enum {string}
 */
export enum ContentViewType {
    Table = 'TABLE',
    Kanban = 'KANBAN',
    UnknownDefaultOpenApi = '11184809'
}


export function instanceOfContentViewType(value: any): boolean {
    for (const key in ContentViewType) {
        if (Object.prototype.hasOwnProperty.call(ContentViewType, key)) {
            if (ContentViewType[key as keyof typeof ContentViewType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ContentViewTypeFromJSON(json: any): ContentViewType {
    return ContentViewTypeFromJSONTyped(json, false);
}

export function ContentViewTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContentViewType {
    return json as ContentViewType;
}

export function ContentViewTypeToJSON(value?: ContentViewType | null): any {
    return value as any;
}

export function ContentViewTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): ContentViewType {
    return value as ContentViewType;
}