// packages/api-client/src/generated/models/ContentSortDirection.ts
/* tslint:disable */
/**
 * Yumpoo API
 * Yumpoo 一期 HTTP API 的唯一静态契约源。  所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共 schema、header、parameter、response 和 golden example；业务 paths 在对应模块 实现时按兼容新增原则加入。
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @enum {string}
 */
export enum ContentSortDirection {
    Asc = 'ASC',
    Desc = 'DESC',
    UnknownDefaultOpenApi = '11184809'
}


export function instanceOfContentSortDirection(value: any): boolean {
    for (const key in ContentSortDirection) {
        if (Object.prototype.hasOwnProperty.call(ContentSortDirection, key)) {
            if (ContentSortDirection[key as keyof typeof ContentSortDirection] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ContentSortDirectionFromJSON(json: any): ContentSortDirection {
    return ContentSortDirectionFromJSONTyped(json, false);
}

export function ContentSortDirectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContentSortDirection {
    return json as ContentSortDirection;
}

export function ContentSortDirectionToJSON(value?: ContentSortDirection | null): any {
    return value as any;
}

export function ContentSortDirectionToJSONTyped(value: any, ignoreDiscriminator: boolean): ContentSortDirection {
    return value as ContentSortDirection;
}