// packages/api-client/src/generated/models/WorkItemPriority.ts
/* tslint:disable */
/**
 * Yumpoo API
 * Yumpoo 一期 HTTP API 的唯一静态契约源。  所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共 schema、header、parameter、response 和 golden example；业务 paths 在对应模块 实现时按兼容新增原则加入。
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @enum {string}
 */
export enum WorkItemPriority {
    Low = 'LOW',
    Medium = 'MEDIUM',
    High = 'HIGH',
    Urgent = 'URGENT',
    UnknownDefaultOpenApi = '11184809'
}


export function instanceOfWorkItemPriority(value: any): boolean {
    for (const key in WorkItemPriority) {
        if (Object.prototype.hasOwnProperty.call(WorkItemPriority, key)) {
            if (WorkItemPriority[key as keyof typeof WorkItemPriority] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WorkItemPriorityFromJSON(json: any): WorkItemPriority {
    return WorkItemPriorityFromJSONTyped(json, false);
}

export function WorkItemPriorityFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkItemPriority {
    return json as WorkItemPriority;
}

export function WorkItemPriorityToJSON(value?: WorkItemPriority | null): any {
    return value as any;
}

export function WorkItemPriorityToJSONTyped(value: any, ignoreDiscriminator: boolean): WorkItemPriority {
    return value as WorkItemPriority;
}