// packages/api-client/src/generated/models/WorkItemRankPlacement.ts
/* tslint:disable */
/**
 * Yumpoo API
 * Yumpoo 一期 HTTP API 的唯一静态契约源。  所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共 schema、header、parameter、response 和 golden example；业务 paths 在对应模块 实现时按兼容新增原则加入。
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @enum {string}
 */
export enum WorkItemRankPlacement {
    Start = 'START',
    Before = 'BEFORE',
    After = 'AFTER',
    End = 'END',
    UnknownDefaultOpenApi = '11184809'
}


export function instanceOfWorkItemRankPlacement(value: any): boolean {
    for (const key in WorkItemRankPlacement) {
        if (Object.prototype.hasOwnProperty.call(WorkItemRankPlacement, key)) {
            if (WorkItemRankPlacement[key as keyof typeof WorkItemRankPlacement] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WorkItemRankPlacementFromJSON(json: any): WorkItemRankPlacement {
    return WorkItemRankPlacementFromJSONTyped(json, false);
}

export function WorkItemRankPlacementFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkItemRankPlacement {
    return json as WorkItemRankPlacement;
}

export function WorkItemRankPlacementToJSON(value?: WorkItemRankPlacement | null): any {
    return value as any;
}

export function WorkItemRankPlacementToJSONTyped(value: any, ignoreDiscriminator: boolean): WorkItemRankPlacement {
    return value as WorkItemRankPlacement;
}